Rename compoennt files to index
