+++ src/components/menu/Menu.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { Button } from '../common/button/Button';
-import { Optional } from '../../shared/ui/optional/Optional';
-import { Select } from '../../shared/ui/select/Select';
+import { Optional } from '../../shared/ui/optional';
+import { Select } from '../../shared/ui/select';
 import { Slider } from '../common/slider/Slider';
 import { MenuEntry } from './components/MenuEntry';
