+++ src/shared/ui/slider/index.tsx
